+++ src/hooks/room-list/fetch-roomlist.ts
@@ -1,15 +1,7 @@
-import { GetUser, GetUserRoomList } from "@/api/user/user-service";
+import { GetUserRoomList } from "@/api/user/user-service";
 import { RoomBasicInfoResponse } from "@/proto-generated/nori/v0/room/room_basic_info_response_pb";
 import { Dispatch, SetStateAction } from 'react';
 
-export const fetchUsername = async (userId: bigint, setUsername: Dispatch<SetStateAction<string>>) => {
-    try {
-        const user = await GetUser(userId);
-        setUsername(user.username);
-    } catch (error) {
-        console.error("Failed to fetch username:", error);
-    }
-};
 
 export const fetchRoomList = async (userId: bigint, setRoomListArray: Dispatch<SetStateAction<RoomBasicInfoResponse[]>>) => {
     try {
@@ -19,4 +11,3 @@ export const fetchRoomList = async (userId: bigint, setRoomListArray: Dispatch<S
         console.error("Failed to fetch room list:", error);
     }
 };
-
