--- src/hooks/room-list/handleIntoRoom.ts
@@ -1,10 +1,0 @@
-// import { useNavigate } from "react-router";
-
-// const navigate = useNavigate();
-// export const handleIntoRoom = (id: bigint) => {
-//     navigate("/roomchat", {
-//       state: {
-//         roomid: id
-//       }
-//     });
-//   };
